feat: 챌린지 일정 가져오는 함수 수정
